refactor: remove unused uikit interaction files from the apps-engine
