+++ packages/apps-engine/src/definition/uikit/index.ts
@@ -1,4 +1,3 @@
-export * from './IUIKitIncomingInteraction';
 export * from './IUIKitSurface';
 export type * from './IUIKitActionHandler';
 export * from './IUIKitInteractionType';
