+++ packages/apps-engine/src/definition/uikit/livechat/index.ts
@@ -1,4 +1,3 @@
 export type * from './IUIKitLivechatActionHandler';
 export type * from './UIKitLivechatIncomingInteractionType';
 export * from './UIKitLivechatInteractionContext';
-export type * from './IUIKitLivechatIncomingInteraction';
