+++ packages/apps/src/server/managers/AppListenerManager.ts
@@ -21,16 +21,8 @@ import { AppInterface, AppMethod } from '@rocket.chat/apps-engine/definition/met
 import type { IRoom, IRoomUserJoinedContext, IRoomUserLeaveContext } from '@rocket.chat/apps-engine/definition/rooms';
 import { RoomType } from '@rocket.chat/apps-engine/definition/rooms';
 import { UIActionButtonContext } from '@rocket.chat/apps-engine/definition/ui';
-import type { IUIKitResponse, IUIKitSurface, UIKitIncomingInteraction } from '@rocket.chat/apps-engine/definition/uikit';
-import { UIKitIncomingInteractionType } from '@rocket.chat/apps-engine/definition/uikit';
-import {
-	isUIKitIncomingInteractionActionButtonMessageBox,
-	isUIKitIncomingInteractionActionButtonMediaCallWidget,
-} from '@rocket.chat/apps-engine/definition/uikit/IUIKitIncomingInteractionActionButton';
-import type {
-	IUIKitLivechatBlockIncomingInteraction,
-	IUIKitLivechatIncomingInteraction,
-} from '@rocket.chat/apps-engine/definition/uikit/livechat';
+import type { IUIKitResponse, IUIKitSurface } from '@rocket.chat/apps-engine/definition/uikit';
+import type { IUIKitLivechatBlockIncomingInteraction } from '@rocket.chat/apps-engine/definition/uikit/livechat';
 import type { IFileUploadInternalContext } from '@rocket.chat/apps-engine/definition/uploads/IFileUploadContext';
 import type { IUser, IUserContext, IUserStatusContext, IUserUpdateContext } from '@rocket.chat/apps-engine/definition/users';
 
@@ -39,6 +31,9 @@ import type { AppManager } from '../AppManager';
 import type { ProxiedApp } from '../ProxiedApp';
 import { Utilities } from '../misc/Utilities';
 import { JSONRPC_METHOD_NOT_FOUND } from '../runtime/deno/AppsEngineDenoRuntime';
+import { UIKitIncomingInteractionType, type UIKitIncomingInteraction } from '../uikit/IUIKitIncomingInteraction';
+import { isUIKitIncomingInteractionActionButtonMessageBox } from '../uikit/UIKitIncomingInteractionActionButton';
+import type { IUIKitLivechatIncomingInteraction } from '../uikit/livechat/IUIKitLivechatIncomingInteraction';
 
 export interface IListenerExecutor {
 	[AppInterface.IPreMessageSentPrevent]: {
